backend/requirements.txt
Flask==3.0.2
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
SQLAlchemy==2.0.26
